Added proper parameters detection (avoids trashing every lang file !) Modified to suit NOCC's lang file indentation Tested and works basically

+++ lang/add_message.sh
@@ -5,13 +5,33 @@
 # Example:  add_message.sh '$htmlNewMessage' 'new message contents'
 #
 # From phpmyadmin http://phpmyadmin.sf.net/ (lem9)
+# Modified for NOCC by Olivier Cahagne
 #
+# TODO: Rewrote to sh only to be compatible with most shells.
+#
+
+if test $# -eq 0
+then
+  echo "Missing parameters (name of var and var value)"
+  echo "Usage: add_message.sh '$<var name>' 'var value'"
+  echo "Example: add_message.sh '\$htmlNewMessage' 'new message contents'"
+  exit 0
+fi
+
+if test $# -eq 1
+then
+  echo "Missing parameter (name of var or var value)"
+  echo "Usage: add_message.sh '$<var name>' 'var value'"
+  echo "Example: add_message.sh '\$htmlNewMessage' 'new message contents'"
+  exit 0
+fi
+
 for file in *.php
 do
         echo $file " "
         grep -v '?>' ${file} > ${file}.new
-        echo "$1=\""  $2  '";  //to translate' >> ${file}.new
-        echo "?>" >> ${file}.new
+        echo "$1 = '"$2"';  //to translate" >> ${file}.new
+        echo '?>' >> ${file}.new
         rm $file
         mv ${file}.new $file
 done
